requirements.txt
Django>=1.9,<1.10
wagtail==1.3.1
wagtail-modeltranslation==0.4.4
django-ckeditor==5.0.3